backend/requirements.txt
# =============================================================================
# SGCG Designer - Backend Dependencies
# Python 3.11+ recommended. Install: pip install -r backend/requirements.txt
# =============================================================================

# --- Core framework ---
# Web framework and WSGI utilities
Flask==3.0.0
Werkzeug==3.0.0

# --- CORS & database ---
# Allow frontend (Vite dev or Hostinger) to call API from different origin
Flask-CORS==4.0.0
# ORM and MySQL driver for templates, projects, work orders
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0

# --- Configuration & mail ---
# Load SECRET_KEY, DATABASE_URL, etc. from .env (never commit .env)
python-dotenv==1.0.0
# Send work order notification emails to admin
Flask-Mail==0.9.1

# --- Production server (Render, Hostinger VPS) ---
# Use: gunicorn -w 4 -b 0.0.0.0:$PORT "backend.app:create_app()"
gunicorn==22.0.0

# --- Image validation (glass type textures: 256x256 PNG/JPG) ---
Pillow==10.4.0

# --- Development & testing ---
pytest==8.3.2
pytest-flask==1.3.0
